move errors into the errors module
