+++ src/gimli/__init__.py
@@ -1,22 +1,7 @@
-from gimli._udunits2 import Unit
-from gimli._udunits2 import UnitEncoding
-from gimli._udunits2 import UnitFormatting
-from gimli._udunits2 import UnitNameError
-from gimli._udunits2 import UnitStatus
-from gimli._udunits2 import UnitSystem
+from gimli._system import UnitSystem
 from gimli._version import __version__
-from gimli.errors import IncompatibleUnitsError
 
 units = UnitSystem()
+del UnitSystem
 
-__all__ = [
-    "__version__",
-    "units",
-    "IncompatibleUnitsError",
-    "Unit",
-    "UnitEncoding",
-    "UnitFormatting",
-    "UnitNameError",
-    "UnitStatus",
-    "UnitSystem",
-]
+__all__ = ["__version__", "units"]
