+++ src/gimli/errors.py
@@ -1,3 +1,9 @@
+from __future__ import annotations
+
+from gimli._constants import STATUS_MESSAGE
+from gimli._constants import UnitStatus
+
+
 class GimliError(Exception):
     pass
 
@@ -9,3 +15,22 @@ class IncompatibleUnitsError(GimliError):
 
     def __str__(self) -> str:
         return f"incompatible units ({self._src!r}, {self._dst!r})"
+
+
+class UnitError(GimliError):
+    def __init__(self, code: UnitStatus, msg: str | None = None):
+        self._code = code
+        self._msg = msg or STATUS_MESSAGE.get(self._code, "Unknown")
+
+    def __str__(self) -> str:
+        return f"{self._msg} (status {self._code})"
+
+
+class UnitNameError(UnitError):
+    def __init__(self, name: str, code: UnitStatus):
+        self._name = name
+        self._code = code
+        self._msg = STATUS_MESSAGE.get(self._code, "Unknown")
+
+    def __str__(self) -> str:
+        return "{0!r}: {self._msg} (status {2})"
